+++ packages/edit/client/Place_edit.js
@@ -1,5 +1,7 @@
 
 Kepler.Place.include({
+
+	geomEdited: false,
 	
 	getOwner: function() {
 		this._dep.depend();
@@ -8,5 +10,25 @@ Kepler.Place.include({
 	
 	isEditable: function() {
 		return !!(this.userId ? (K.Profile.id === this.userId) : false);
+	},
+
+	setGeometry(geoj) {
+		this._dep.depend();
+		
+		console.log('setGeometry',geoj, this)
+
+		this.geometry = geoj.geometry;
+		
+		if(this.geom) {
+			this.geom.remove();
+			delete this.geom;
+		}
+		
+		this.buildGeometry();//fill this.geom
+		//this.update();
+
+		this.geomEdited = true;
+
+		return this;
 	}
 });