+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - plugin import support geojson LineString centroid
 - plugin categories new template panelPlaceEdit_cats_latest
 - plugin categories fix remove cat button
+- plugin edit split template panelPlaceEdit in panelPlaceEdit_edit_map
 
 ## v1.6.6
 - core support Instagram OAuth signup